added service implementation.  Imported new dependencies.  Previous package was incorrect
