+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/DemoApplication.java
@@ -1,0 +1,13 @@
+package com.example.springbootmonolith;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DemoApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(DemoApplication.class, args);
+	}
+
+}
