+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/Repository/CommentRepository.java
@@ -1,0 +1,10 @@
+package com.example.springbootmonolith.Repository;
+
+import com.example.springbootmonolith.model.Comment;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CommentRepository extends CrudRepository<Comment, Long> {
+
+}
