+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/Repository/UserRepository.java
@@ -1,0 +1,9 @@
+package com.example.springbootmonolith.Repository;
+
+import com.example.springbootmonolith.model.User;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends CrudRepository<User, Long> {
+}
