+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/Service/CommentService.java
@@ -1,0 +1,10 @@
+package com.example.springbootmonolith.Service;
+
+import com.example.springbootmonolith.model.Comment;
+
+public interface CommentService {
+
+    public Comment createComment (Comment comment);
+
+    public Iterable<Comment> listComments();
+}
