+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/Service/CommentServiceImpl.java
@@ -1,0 +1,22 @@
+package com.example.springbootmonolith.Service;
+
+import com.example.springbootmonolith.Repository.CommentRepository;
+import com.example.springbootmonolith.model.Comment;
+import org.springframework.beans.factory.annotation.Autowired;
+
+public class CommentServiceImpl implements CommentService {
+
+    @Autowired
+    CommentRepository commentRepository;
+
+    @Override
+    public Comment createComment(Comment comment) {
+        return commentRepository.save(comment);
+    }
+
+    @Override
+    public Iterable<Comment> listComments(){
+        return commentRepository.findAll();
+    }
+
+}
