+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/Service/PostService.java
@@ -1,0 +1,11 @@
+package com.example.springbootmonolith.Service;
+
+import com.example.springbootmonolith.model.Post;
+
+public interface PostService {
+
+    public Post createPost(Post post);
+
+    public Iterable<Post> listPosts();
+
+}
