+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/Service/UserProfileService.java
@@ -1,0 +1,10 @@
+package com.example.springbootmonolith.Service;
+
+import com.example.springbootmonolith.model.UserProfile;
+
+public interface UserProfileService {
+
+    public UserProfile createUserProfile(String username, UserProfile newProfile);
+
+    public UserProfile getUserProfile(String username);
+}
