+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/Service/UserService.java
@@ -1,0 +1,19 @@
+package com.example.springbootmonolith.Service;
+
+import com.example.springbootmonolith.model.User;
+
+public interface UserService {
+
+    public String login (User user);
+
+    public User createUser (User newUser);
+
+    public User addPost(String username, long postId);
+
+    public User addComment(String username,long commentId);
+
+    public Iterable<User> listUsers();
+
+    public User getUser(String username);
+
+}
