+++ Spring-boot-monolith/src/main/java/com/example/springbootmonolith/model/Comment.java
@@ -1,0 +1,54 @@
+package com.example.springbootmonolith.model;
+
+import javax.persistence.*;
+
+
+@Entity
+@Table(name = "comment")
+public class Comment {
+    @Id
+    @Column
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    @Column
+    private String title;
+
+    @Column
+    private String body;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    private Comment() {}
+
+    public User getUser(){return user;}
+
+    public void setUser(User user){this.user = user;}
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getBody() {
+        return body;
+    }
+
+    public void setBody(String body) {
+        this.body = body;
+    }
+
+}
