+++ Spring-boot-monolith/src/main/resources/application-dev.properties
@@ -1,0 +1,16 @@
+spring.datasource.initialization-mode=always
+spring.datasource.platform=postgres
+spring.datasource.url=jdbc:postgresql://localhost:5432/penny
+
+spring.datasource.username=tom
+spring.datasource.password=tom123
+
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=none
+
+spring.jpa.generate-ddl=true
+
+spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
+
+# CREATE DATABASE testing;
+# GRANT ALL PRIVILEGES ON DATABASE testing to tom;
