+++ Spring-boot-monolith/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+spring.profiles.active=dev
+jwt.secret=pancakes
