+++ Spring-boot-monolith/src/test/java/com/example/springbootmonolith/DemoApplicationTests.java
@@ -1,0 +1,16 @@
+package com.example.springbootmonolith;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class DemoApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
