--- apiPROJECT/src/main/java/com/example/apiPROJECT/ApiProjectApplication.java
@@ -1,13 +1,0 @@
-package com.example.apiPROJECT;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ApiProjectApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ApiProjectApplication.class, args);
-	}
-
-}
