--- apiPROJECT/src/main/java/com/example/apiPROJECT/repository/PostRepository.java
@@ -1,11 +1,0 @@
-package com.example.apiPROJECT.repository;
-
-import com.example.apiPROJECT.model.Post;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PostRepository extends CrudRepository <Post, Long>  {
-
-    public Post findByPostId(Post id);
-}
