--- apiPROJECT/src/main/java/com/example/apiPROJECT/repository/UserProfileRepository.java
@@ -1,11 +1,0 @@
-package com.example.apiPROJECT.repository;
-
-import com.example.apiPROJECT.model.UserProfile;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserProfileRepository extends CrudRepository <UserProfile, Long>  {
-
-    public UserProfile findByProfileId(UserProfile id);
-}
