--- apiPROJECT/src/main/resources/application.properties
@@ -1,12 +1,0 @@
-spring.datasource.initialization-mode=always
-spring.datasource.platform=postgres
-spring.datasource.url=jdbc:postgresql://localhost:5432/penny
-spring.jpa.show-sql=true
-spring.jpa.hibernate.ddl-auto=create
-
-server.port=8080
-spring.datasource.username=tom
-spring.datasource.password=tom123
-
-spring.main.allow-bean-definition-overriding=true
-#spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
