--- apiPROJECT/src/test/java/com/example/apiPROJECT/ApiProjectApplicationTests.java
@@ -1,16 +1,0 @@
-package com.example.apiPROJECT;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class ApiProjectApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
